Refactor some nampspaces to make it easier to navigate API models
