+++ src/Altinn.Profile/Models/AddressVerification/AddressVerificationRequest.cs
@@ -1,14 +1,14 @@
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
 using Altinn.Profile.Core.AddressVerifications.Models;
 
-namespace Altinn.Profile.Models
+namespace Altinn.Profile.Models.AddressVerification
 {
     /// <summary>
-    /// Represents a request to resend a verification code for a given address (email or phone number)
+    /// Represents a request to verify an address, such as an email or phone number, using a verification code.
     /// </summary>
-    public class AddressCodeResendRequest
+    public class AddressVerificationRequest
     {
         /// <summary>
         /// Gets or sets the address to verify, either an email or a phone number.
@@ -26,5 +26,14 @@ namespace Altinn.Profile.Models
         [JsonRequired]
         [JsonConverter(typeof(JsonStringEnumConverter))]
         public required AddressType? Type { get; init; }
+
+        /// <summary>
+        /// Gets or sets the verification code for the address.
+        /// </summary>
+        [Required]
+        [JsonRequired]
+        [StringLength(6, MinimumLength = 6)]
+        [RegularExpression(@"^\d{6}$")]
+        public required string VerificationCode { get; init; }
     }
 }
