+++ src/Altinn.Profile/Models/OrganizationNotificationAddresses/NotificationAddressResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace Altinn.Profile.Models
+﻿namespace Altinn.Profile.Models.OrganizationNotificationAddresses
 {
     /// <summary>
     /// Represents a notification address
