+++ src/Altinn.Profile/Models/OrganizationNotificationAddresses/OrganizationResponse.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Altinn.Profile.Models
+namespace Altinn.Profile.Models.OrganizationNotificationAddresses
 {    
     /// <summary>
      /// Represents a on organization with  notification addresses
