+++ src/Altinn.Profile/Models/ProfessionalNotificationSettings/NotificationSettingsPatchRequest.cs
