+++ src/Altinn.Profile/Models/ProfessionalNotificationSettings/NotificationSettingsRequest.cs
@@ -1,11 +1,13 @@
 ﻿#nullable enable
 
-namespace Altinn.Profile.Models
+using Altinn;
+
+namespace Altinn.Profile.Models.ProfessionalNotificationSettings
 {
     /// <summary>
     /// Request model for the professional notification address for an organization, also called personal notification address.
     /// </summary>
-    public class NotificationSettingsRequest : ProfessionalNotificationAddress
+    public class NotificationSettingsRequest : ProfessionalNotificationSettings
     {
     }
 }
