+++ test/Altinn.Profile.Tests/IntegrationTests/API/Controllers/NotificationsSettingsControllerTests.cs
@@ -13,6 +13,7 @@ using Altinn.Profile.Core.ProfessionalNotificationAddresses;
 using Altinn.Profile.Core.User.ProfileSettings;
 using Altinn.Profile.Core.Utils;
 using Altinn.Profile.Models;
+using Altinn.Profile.Models.ProfessionalNotificationSettings;
 using Altinn.Profile.Tests.IntegrationTests.Utils;
 
 using Microsoft.AspNetCore.Http;
